+++ tests/std/tests/VSO_0000000_vector_algorithms_floats/test.cpp
@@ -20,48 +20,9 @@
 #endif // _HAS_CXX20
 
 #include "test_min_max_element_support.hpp"
+#include "test_vector_algorithms_support.hpp"
 
 using namespace std;
-
-#pragma warning(disable : 4984) // 'if constexpr' is a C++17 language extension
-#ifdef __clang__
-#pragma clang diagnostic ignored "-Wc++17-extensions" // constexpr if is a C++17 extension
-#endif // __clang__
-
-#if (defined(_M_IX86) || defined(_M_X64)) && !defined(_M_CEE_PURE)
-extern "C" long __isa_enabled;
-
-void disable_instructions(ISA_AVAILABILITY isa) {
-    __isa_enabled &= ~(1UL << static_cast<unsigned long>(isa));
-}
-#endif // (defined(_M_IX86) || defined(_M_X64)) && !defined(_M_CEE_PURE)
-
-void initialize_randomness(mt19937_64& gen) {
-    constexpr size_t n = mt19937_64::state_size;
-    constexpr size_t w = mt19937_64::word_size;
-    static_assert(w % 32 == 0, "w should be evenly divisible by 32");
-    constexpr size_t k = w / 32;
-
-    vector<uint32_t> vec(n * k);
-
-    random_device rd;
-    generate(vec.begin(), vec.end(), ref(rd));
-
-    printf("This is a randomized test.\n");
-    printf("DO NOT IGNORE/RERUN ANY FAILURES.\n");
-    printf("You must report them to the STL maintainers.\n\n");
-
-    printf("Seed vector: ");
-    for (const auto& e : vec) {
-        printf("%u,", e);
-    }
-    printf("\n");
-
-    seed_seq seq(vec.cbegin(), vec.cend());
-    gen.seed(seq);
-}
-
-constexpr size_t dataCount = 1024;
 
 template <class T>
 void test_min_max_element_floating_with_values(mt19937_64& gen, const std::vector<T>& input_of_input) {
@@ -150,17 +111,5 @@ void test_vector_algorithms(mt19937_64& gen) {
 }
 
 int main() {
-    mt19937_64 gen;
-    initialize_randomness(gen);
-
-    test_vector_algorithms(gen);
-#ifndef _M_CEE_PURE
-#if defined(_M_IX86) || defined(_M_X64)
-    disable_instructions(__ISA_AVAILABLE_AVX2);
-    test_vector_algorithms(gen);
-
-    disable_instructions(__ISA_AVAILABLE_SSE42);
-    test_vector_algorithms(gen);
-#endif // defined(_M_IX86) || defined(_M_X64)
-#endif // _M_CEE_PURE
+    run_randomized_tests_with_different_isa_levels(test_vector_algorithms);
 }
